add cookie and signout api

+++ admin/src/utils/actions/auth.js
@@ -53,14 +53,23 @@ export const isUserLoggedIn = () => {
     }
 }
 export const signout = (user) =>{
-    console.log(user)
     return async (dispatch)=>{
-        localStorage.clear()
-        dispatch({
-            type:authConstant.LOGOUT_REQUEST,
-            payload:{
-                error: "failed to login"
-            }
-        })
+       dispatch({type:authConstant.LOGOUT_REQUEST})
+      const res =  await axios.post(`/admin/signout`)
+        if(res.status===200){
+            localStorage.clear()
+            dispatch({
+                type:authConstant.LOGOUT_SUCCESS
+            })
+        }else {
+            dispatch({
+                type:authConstant.LOGOUT_FAILURE,
+                payload:{
+                    error:res.data.error
+                }
+            })
+        }
+
+
     }
 }
